+++ index.html
@@ -247,7 +247,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="./js/main.js" defer></script>
+  <script type="module" src="./js/main.js" defer></script>
 </body>
 
 </html>
