+++ js/comment-length.js
@@ -1,0 +1,8 @@
+const maxStringLength = function (sampleLength, maxLength) {
+  if (sampleLength.length > maxLength) {
+    return false;
+  }
+  return true;
+}
+
+export { maxStringLength };
