+++ js/user-names.js
@@ -1,0 +1,7 @@
+const names = ['Иван', 'Шерлок', 'Яков', 'Макар',
+  'Доминик', 'Ян', 'Шарль', 'Леонард', 'Феликс',
+  'Олег', 'Донат', 'Карен', 'Юрий', 'Вячеслав',
+  'Степан', 'Оскар', 'Ярослав', 'Яков', 'Чарльз',
+  'Тит', 'Чарльз', 'Камиль', 'Эрик', 'Даниил', 'Адам'];
+
+export { names };
